Examples: 将各C++示例从x86切换到x64并修正编译问题。

+++ Examples/Communication/CPP/Src/Client/Clients/TcpClient.h
@@ -125,7 +125,7 @@ namespace Client
             {
                 while (true)
                 {
-                    auto r = VirtualTransportClient->Handle(Count);
+                    auto r = VirtualTransportClient->Handle(static_cast<int>(Count));
                     if (r->OnContinue())
                     {
                         break;
